+++ packages/amplify-category-api/src/__tests__/provider-utils/awscloudformation/utils/import-rds-utils/globalAmplifyInputs.test.ts
@@ -32,7 +32,7 @@ describe('Amplify Input read/write from schema', () => {
   });
 
   it('constructs valid default input parameters for MySQL datasource with global auth rule', async () => {
-    const expectedGraphQLInputString = `input Amplify {
+    const expectedGraphQLInputString = `input AMPLIFY {
       engine: String = \"mysql\"
       globalAuthRule: AuthRule = { allow: public } # This "input" configures a global authorization rule to enable public access to all models in this schema. Learn more about authorization rules here:https://docs.amplify.aws/cli/graphql/authorization-rules 
     }`;
@@ -48,7 +48,7 @@ describe('Amplify Input read/write from schema', () => {
       database: 'mockdatabase',
     };
 
-    const mockInputSchema = `input Amplify {
+    const mockInputSchema = `input AMPLIFY {
       engine: String = \"${mockValidInputs.engine}\"  
       globalAuthRule: AuthRule = { allow: public } # This "input" configures a global authorization rule to enable public access to all models in this schema. Learn more about authorization rules here:https://docs.amplify.aws/cli/graphql/authorization-rules 
     }`;
@@ -68,7 +68,7 @@ describe('Amplify Input read/write from schema', () => {
       database: 'mockdatabase',
     };
 
-    const mockInputSchema = `input Amplify {
+    const mockInputSchema = `input AMPLIFY {
       engine: String = \"${mockValidInputs.engine}\" 
       globalAuthRule: AuthRule = { allow: public } # This "input" configures a global authorization rule to enable public access to all models in this schema. Learn more about authorization rules here:https://docs.amplify.aws/cli/graphql/authorization-rules 
     }`;