+++ packages/amplify-category-api/src/provider-utils/awscloudformation/service-walkthroughs/import-appsync-api-walkthrough.ts
@@ -1,4 +1,5 @@
 import * as path from 'path';
+import * as fs from 'fs-extra';
 import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { prompter, printer, integer } from '@aws-amplify/amplify-prompts';
 import {
@@ -7,23 +8,21 @@ import {
   ImportedRDSType,
   ImportedDataSourceConfig,
 } from '@aws-amplify/graphql-transformer-core';
+import { storeConnectionSecrets, getSecretsKey, getExistingConnectionSecrets } from '../utils/rds-resources/database-resources';
+import { parseDatabaseUrl } from '../utils/database-url';
 import { RDS_SCHEMA_FILE_NAME } from '@aws-amplify/graphql-transformer-core';
-import * as fs from 'fs-extra';
-import { serviceMetadataFor } from '../utils/dynamic-imports';
-import { getCfnApiArtifactHandler } from '../cfn-api-artifact-handler';
-import { serviceWalkthroughResultToAddApiRequest } from '../utils/service-walkthrough-result-to-add-api-request';
-import { writeSchemaFile } from '../utils/graphql-schema-utils';
-import { PREVIEW_BANNER, category } from '../../../category-constants';
-import { storeConnectionSecrets, testDatabaseConnection, getSecretsKey } from '../utils/rds-secrets/database-secrets';
 import { constructDefaultGlobalAmplifyInput } from '../utils/rds-input-utils';
 import { getAPIResourceDir } from '../utils/amplify-meta-utils';
 import { getAppSyncAPINames } from '../utils/amplify-meta-utils';
+import { writeSchemaFile } from '../utils/graphql-schema-utils';
 import { serviceApiInputWalkthrough } from './appSync-walkthrough';
+import { serviceMetadataFor } from '../utils/dynamic-imports';
+import { serviceWalkthroughResultToAddApiRequest } from '../utils/service-walkthrough-result-to-add-api-request';
+import { getCfnApiArtifactHandler } from '../cfn-api-artifact-handler';
 
 const service = 'AppSync';
 
 export const importAppSyncAPIWalkthrough = async (context: $TSContext): Promise<ImportAppSyncAPIInputs> => {
-  printer.warn(PREVIEW_BANNER);
   let apiName: string;
   const existingAPIs = getAppSyncAPINames();
   if (existingAPIs?.length > 0) {
@@ -51,10 +50,9 @@ export const importAppSyncAPIWalkthrough = async (context: $TSContext): Promise<
 
   await writeDefaultGraphQLSchema(context, pathToSchemaFile, databaseConfig);
   await storeConnectionSecrets(context, databaseConfig, apiName, secretsKey);
-  await testDatabaseConnection(databaseConfig);
 
   return {
-    apiName: apiName,
+    apiName,
     dataSourceConfig: databaseConfig,
   };
 };
@@ -63,7 +61,7 @@ export const writeDefaultGraphQLSchema = async (
   context: $TSContext,
   pathToSchemaFile: string,
   databaseConfig: ImportedDataSourceConfig,
-) => {
+): Promise<void> => {
   const dataSourceType = databaseConfig?.engine;
   if (Object.values(ImportedRDSType).includes(dataSourceType)) {
     const globalAmplifyInputTemplate = await constructDefaultGlobalAmplifyInput(context, databaseConfig.engine);
@@ -73,32 +71,50 @@ export const writeDefaultGraphQLSchema = async (
   }
 };
 
-export const databaseConfigurationInputWalkthrough = async (
-  engine: ImportedDataSourceType,
-  database?: string,
-): Promise<ImportedDataSourceConfig> => {
-  const databaseName = database || (await prompter.input(`Enter the name of the ${formatEngineName(engine)} database to import:`));
-  const host = await prompter.input(`Enter the host for ${databaseName} database:`);
-  const port = await prompter.input<'one', number>(`Enter the port for ${databaseName} database:`, {
-    transform: (input) => Number.parseInt(input, 10),
-    validate: integer(),
-    initial: 3306,
-  });
+export const databaseConfigurationInputWalkthrough = async (engine: ImportedDataSourceType): Promise<ImportedDataSourceConfig> => {
+  printer.info('Please provide the following database connection information:');
+  const url = await prompter.input('Enter the database url or host name:');
+
+  let isValidUrl = true;
+  const parsedDatabaseUrl = parseDatabaseUrl(url);
+  let { host, port, database, username, password } = parsedDatabaseUrl;
+
+  if (!host) {
+    isValidUrl = false;
+    host = url;
+  }
+  if (!isValidUrl || !port) {
+    port = await prompter.input<'one', number>('Enter the port number:', {
+      transform: (input) => Number.parseInt(input, 10),
+      validate: integer(),
+      initial: 3306,
+    });
+  }
+
   // Get the database user credentials
-  const username = await prompter.input(`Enter the username for ${databaseName} database user:`);
-  const password = await prompter.input(`Enter the password for ${databaseName} database user:`, { hidden: true });
+  if (!isValidUrl || !username) {
+    username = await prompter.input('Enter the username:');
+  }
+
+  if (!isValidUrl || !password) {
+    password = await prompter.input('Enter the password:', { hidden: true });
+  }
+
+  if (!isValidUrl || !database) {
+    database = await prompter.input('Enter the database name:');
+  }
 
   return {
-    engine: engine,
-    database: databaseName,
-    host: host,
-    port: port,
-    username: username,
-    password: password,
+    engine,
+    database,
+    host,
+    port,
+    username,
+    password,
   };
 };
 
-export const formatEngineName = (engine: ImportedDataSourceType) => {
+export const formatEngineName = (engine: ImportedDataSourceType): string => {
   switch (engine) {
     case ImportedRDSType.MYSQL:
       return 'MySQL';