+++ packages/amplify-graphql-model-transformer/src/__tests__/model-transformer.test.ts
@@ -1487,4 +1487,29 @@ describe('ModelTransformer: ', () => {
     const updateTodoIdField = getFieldOnInputType(updateTodoInput!, 'id');
     expect(updateTodoIdField.type.kind).toBe('NonNullType');
   });
+
+  it('should successfully transform simple rds valid schema', async () => {
+    const validSchema = `
+      type Post @model {
+          id: ID!
+          title: String!
+      }
+    `;
+
+    const transformer = new GraphQLTransform({
+      transformers: [new ModelTransformer()],
+    });
+    const modelToDatasourceMap = new Map<string, DatasourceType>();
+    modelToDatasourceMap.set('Post', {
+      dbType: 'MySQL',
+      provisionDB: false,
+    });
+    const out = transformer.transform(validSchema, {
+      modelToDatasourceMap,
+    });
+    expect(out).toBeDefined();
+
+    validateModelSchema(parse(out.schema));
+    parse(out.schema);
+  });
 });