+++ packages/amplify-migration-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.31](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.30...amplify-category-api-migration-tests@5.0.31) (2023-07-21)
+
+**Note:** Version bump only for package amplify-category-api-migration-tests
+
 ## [5.0.30](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.29...amplify-category-api-migration-tests@5.0.30) (2023-07-17)
 
 **Note:** Version bump only for package amplify-category-api-migration-tests